experiments/audio-engine/audio-automation/requirements.txt
yt-dlp
streamlink
demucs
basic-pitch
pychord
pyacoustid
beets
PyYAML
tqdm
rich
pytest
faster-whisper
whisperx
torch
torchaudio
webvtt-py
pysrt
Flask
